chore: adjusted ios notifications handling
